Split graph tool settings module
