+++ tests/tool/graph_settings_test.py
@@ -1,0 +1,8 @@
+from avalan.tool.graph_settings import GraphToolSettings
+
+
+def test_graph_tool_settings_dict_mapping() -> None:
+    settings = GraphToolSettings(file="/tmp/chart.png")
+
+    assert settings.file == "/tmp/chart.png"
+    assert settings["file"] == "/tmp/chart.png"
